Use different ButtonStyleData for DropdownButtonFormField2
